refactor: use useQuery for state management in settings
