+++ admin/src/components/settings/index.ts
@@ -1,2 +1,3 @@
 export * from './SettingsToggleField';
 export * from './SettingsTextField';
+export * from './SettingsPage';
